// jobup-frontend/src/app/generated-sources/openapi/model/workerCreateDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * DTO used to create a new worker
 */
export interface WorkerCreateDto { 
    jobType?: string;
    phoneNumber?: string;
    location?: string;
    description?: string;
    /**
     * User ID who becomes a worker
     */
    userId?: string;
}